Feature: Make booking lead days configurable
This commit introduces a new API endpoint to fetch the maximum number of days in the future a booking can be made from the booking settings.

The datepicker on the resource availability page now uses this value to dynamically set the maximum selectable date. If the value is not set, it defaults to 365 days.
